Revert "Review mario 2 (#5)"
This reverts commit cbd7390a1b2da3f53a00daac5f72c0205e082b57.
