--- project/Versions.scala
@@ -1,8 +1,0 @@
-object Versions {
-
-  val spark = "3.5.1"
-  val delta = "3.2.0"
-  val scalaTest = "3.2.19"
-  val sparkFastTests = "1.3.0"
-
-}
